v0.2.0 release

+++ updater/src/net/lecousin/dataorganizer/updater/Installer.java
@@ -9,11 +9,13 @@ import net.lecousin.framework.progress.WorkProgress;
 public class Installer {
 
 	public static boolean installApplication(File appDir, File tmpAppDir, WorkProgress progress, int work) {
+		if (tmpAppDir.listFiles().length == 0) return true; // handle less_than case (empty app to avoid 0.1 bug)
 		FileSystemUtil.deleteDirectory(new File(appDir, "configuration"));
 		FileSystemUtil.deleteDirectory(new File(appDir, "plugins"));
 		for (File file : appDir.listFiles()) {
 			if (file.isDirectory()) continue;
 			if (file.getName().equalsIgnoreCase("DataOrganizer.xml")) continue;
+			if (file.getName().equalsIgnoreCase("DataOrganizer.ini")) continue;
 			if (file.getName().equalsIgnoreCase("recentWorkspaces.xml")) continue;
 			file.delete();
 		}
